разбила код на отдельные модули, вынесла каждый из экранов в отдельный модуль, а также добавила модуль с пользовательскими функциями создания элемента и его отрисовки, добавила логику переключения между экранами
