+++ js/game-2.js
@@ -1,0 +1,72 @@
+import {getElement, render} from './utils';
+import {showGame3} from './game-3';
+import {showGreeting} from './greeting';
+
+const game2Template = `<header class="header">
+  <div class="header__back">
+    <button class="back">
+      <img src="img/arrow_left.svg" width="45" height="45" alt="Back">
+      <img src="img/logo_small.svg" width="101" height="44">
+    </button>
+  </div>
+  <h1 class="game__timer">NN</h1>
+  <div class="game__lives">
+    <img src="img/heart__empty.svg" class="game__heart" alt="Life" width="32" height="32">
+    <img src="img/heart__full.svg" class="game__heart" alt="Life" width="32" height="32">
+    <img src="img/heart__full.svg" class="game__heart" alt="Life" width="32" height="32">
+  </div>
+</header>
+<div class="game">
+  <p class="game__task">Угадай, фото или рисунок?</p>
+  <form class="game__content  game__content--wide">
+    <div class="game__option">
+      <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
+      <label class="game__answer  game__answer--photo">
+        <input name="question1" type="radio" value="photo">
+        <span>Фото</span>
+      </label>
+      <label class="game__answer  game__answer--wide  game__answer--paint">
+        <input name="question1" type="radio" value="paint">
+        <span>Рисунок</span>
+      </label>
+    </div>
+  </form>
+  <div class="stats">
+    <ul class="stats">
+      <li class="stats__result stats__result--wrong"></li>
+      <li class="stats__result stats__result--slow"></li>
+      <li class="stats__result stats__result--fast"></li>
+      <li class="stats__result stats__result--correct"></li>
+      <li class="stats__result stats__result--wrong"></li>
+      <li class="stats__result stats__result--unknown"></li>
+      <li class="stats__result stats__result--slow"></li>
+      <li class="stats__result stats__result--unknown"></li>
+      <li class="stats__result stats__result--fast"></li>
+      <li class="stats__result stats__result--unknown"></li>
+    </ul>
+  </div>
+</div>
+<footer class="footer">
+  <a href="https://htmlacademy.ru" class="social-link social-link--academy">HTML Academy</a>
+  <span class="footer__made-in">
+    Сделано в
+    <a href="https://htmlacademy.ru" class="footer__link">HTML Academy</a>
+    &copy; 2016
+  </span>
+  <div class="footer__social-links">
+    <a href="https://twitter.com/htmlacademy_ru" class="social-link  social-link--tw">Твиттер</a>
+    <a href="https://www.instagram.com/htmlacademy/" class="social-link  social-link--ins">Инстаграм</a>
+    <a href="https://www.facebook.com/htmlacademy" class="social-link  social-link--fb">Фэйсбук</a>
+    <a href="https://vk.com/htmlacademy" class="social-link  social-link--vk">Вконтакте</a>
+  </div>
+</footer>`;
+
+export const showGame2 = () => {
+  const game2 = getElement(game2Template);
+
+  const content = game2.querySelector(`.game__content`);
+
+  content.addEventListener(`input`, showGame3);
+
+  render(game2, showGreeting);
+};
